Add PEXE loading test.
Also fix bug in char6 abbrev op: it should return an array.

The pexe is loaded and compared to JSON generated by another program. As
a result, the data has to be munged. It was simplest to do that in the
test code, but it might be nicer to change the golden file.